+++ src/LibraryListPanel.js
@@ -1,7 +1,35 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import LibraryList from './LibraryList';
+const style = {
+  marginLeft: 10
+};
+function Icon({ status }) {
+  switch (status) {
+    case 'loaded':
+      return (
+        <i
+          className="fa fa-check"
+          style={{
+            ...style,
+            color: 'green'
+          }}
+        />
+      );
+    case 'error':
+      return (
+        <i
+          className="fa fa-times"
+          style={{
+            ...style,
+            color: 'red'
+          }}
+        />
+      );
+    default:
+      return <i className="fa fa-spinner fa-pulse" style={style} />;
+  }
+}
 
 const LibraryListPanel = ({ libraryListIsOpen, libraries }) => {
   if (!libraryListIsOpen) return <div />;
@@ -9,7 +37,14 @@ const LibraryListPanel = ({ libraryListIsOpen, libraries }) => {
   return (
     <div style={{ padding: 10 }}>
       <strong>External Libraries</strong>
-      <LibraryList libraries={libraries} />
+      <ul style={{ margin: 5 }}>
+        {libraries.map(({ name, url, status }) => (
+          <li key={url}>
+            <a href={url}>{name || url}</a>
+            <Icon status={status} />
+          </li>
+        ))}
+      </ul>
     </div>
   );
 };
